refactor: improve async/sync pagination and rename pagination methods
