+++ sinch/core/types.py
@@ -1,0 +1,4 @@
+from typing import TypeVar
+from pydantic import BaseModel
+
+BM = TypeVar("BM", bound=BaseModel)
