+++ sinch/domains/numbers/endpoints/available/activate_number_endpoint.py
@@ -1,3 +1,4 @@
+import json
 from sinch.core.enums import HTTPAuthentication, HTTPMethods
 from sinch.core.models.http_response import HTTPResponse
 from sinch.domains.numbers.endpoints.numbers_endpoint import NumbersEndpoint
@@ -15,6 +16,11 @@ class ActivateNumberEndpoint(NumbersEndpoint):
 
     def __init__(self, project_id: str, request_data: ActivateNumberRequest):
         super(ActivateNumberEndpoint, self).__init__(project_id, request_data)
+
+    def request_body(self) -> str:
+        # Convert the request data to a dictionary and remove None values
+        request_data = self.request_data.model_dump(by_alias=True, exclude_none=True)
+        return json.dumps(request_data)
 
     def handle_response(self, response: HTTPResponse) -> ActivateNumberResponse:
         try:
