+++ sinch/domains/numbers/endpoints/numbers_endpoint.py
@@ -1,10 +1,10 @@
-import json
 from abc import ABC
 from typing import Type
 from sinch.core.models.http_response import HTTPResponse
 from sinch.core.endpoint import HTTPEndpoint
+from sinch.core.types import BM
 from sinch.domains.numbers.exceptions import NumbersException
-from sinch.domains.numbers.models.numbers import BM, NotFoundError
+from sinch.domains.numbers.models.numbers import NotFoundError
 
 
 class NumbersEndpoint(HTTPEndpoint, ABC):
@@ -38,9 +38,7 @@ class NumbersEndpoint(HTTPEndpoint, ABC):
         Returns:
             str: The request body as a JSON string.
         """
-        # Convert the request data to a dictionary and remove None values
-        request_data = self.request_data.model_dump(by_alias=True, exclude_none=True)
-        return json.dumps(request_data)
+        return ""
 
     def process_response_model(self, response_body: dict, response_model: Type[BM]) -> BM:
         """
