+++ tests/unit/test_pagination.py
@@ -166,33 +166,6 @@ def test_page_token_iterator_sync_using_manual_pagination(
     assert page_counter == 3
 
 
-def test_page_token_iterator_sync_using_auto_pagination(
-    token_based_pagination_request_data,
-    first_token_based_pagination_response,
-    second_token_based_pagination_response
-):
-    endpoint = Mock()
-    endpoint.request_data = token_based_pagination_request_data
-    sinch_client = Mock()
-
-    sinch_client.configuration.transport.request.side_effect = [
-        first_token_based_pagination_response,
-        second_token_based_pagination_response
-    ]
-    token_based_paginator = TokenBasedPaginator._initialize(
-        sinch=sinch_client,
-        endpoint=endpoint
-    )
-    assert token_based_paginator
-
-    page_counter = 0
-    for page in token_based_paginator.auto_paging_iter():
-        page_counter += 1
-        assert isinstance(page, TokenBasedPaginator)
-
-    assert page_counter == 2
-
-
 def test_page_token_iterator_numbers_sync_using_auto_pagination_expects_iter(token_based_pagination_request_data,
                                                                              mock_pagination_active_number_responses):
     """ Test that the pagination iterates correctly through multiple items. """
@@ -230,7 +203,7 @@ def test_page_token_iterator_sync_using_list_expects_correct_metadata(token_base
     )
     assert token_based_paginator
 
-    list_response = token_based_paginator.list()
+    list_response = token_based_paginator.get_content()
 
     page_counter = 0
     reached_last_page = False
@@ -275,35 +248,6 @@ async def test_page_token_iterator_async_using_manual_pagination(
     assert page_counter == 3
 
 
-async def test_page_token_iterator_async_using_auto_pagination(
-    token_based_pagination_request_data,
-    first_token_based_pagination_response,
-    second_token_based_pagination_response,
-    third_token_based_pagination_response
-):
-    endpoint = Mock()
-    endpoint.request_data = token_based_pagination_request_data
-    sinch_client = AsyncMock()
-
-    sinch_client.configuration.transport.request.side_effect = [
-        first_token_based_pagination_response,
-        second_token_based_pagination_response,
-        third_token_based_pagination_response
-    ]
-    token_based_paginator = await AsyncTokenBasedPaginator._initialize(
-        sinch=sinch_client,
-        endpoint=endpoint
-    )
-    assert token_based_paginator
-
-    page_counter = 0
-    async for page in token_based_paginator.auto_paging_iter():
-        page_counter += 1
-        assert isinstance(page, AsyncTokenBasedPaginator)
-
-    assert page_counter == 3
-
-
 async def test_page_token_iterator_async_using_list_expects_correct_metadata(
     token_based_pagination_request_data,
     mock_pagination_active_number_responses
@@ -322,19 +266,19 @@ async def test_page_token_iterator_async_using_list_expects_correct_metadata(
     )
     assert async_token_based_paginator
 
-    list_response = await async_token_based_paginator.list()
+    list_response = await async_token_based_paginator.get_content()
 
     page_counter = 0
     reached_last_page = False
 
     while not reached_last_page:
-        page_counter += 1
         if list_response.has_next_page:
             list_response = await list_response.next_page()
+            page_counter += 1
         else:
             reached_last_page = True
 
-    assert page_counter == 3
+    assert page_counter == 2
 
 
 async def test_page_token_iterator_numbers_async_using_auto_pagination_expects_iter(
